Разил код на модули
